backend/requirements.txt
Django~=4.1
psycopg2>=2.8
djangorestframework == 3.14.0
django-cors-headers == 3.9.0
djangorestframework_simplejwt == 5.2.2